64-bit TA debugging
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@14921 9d14d920-75d5-0310-8bd7-a52572981aee

+++ agents/linux/ftp_routines.c
@@ -424,7 +424,7 @@ read_test(void *arg)
 {
     char    buf[FTP_GET_BULK];
     int     len = 0;
-    int     sock = (int)arg;
+    int     sock = (int)(long int)arg;
     fd_set  rd;
     int     ret, l;
